+++ MBBSEmu.Tests/CPU/SUB_Tests.cs
@@ -493,10 +493,10 @@ namespace MBBSEmu.Tests.CPU
             Assert.Equal(expectedValue, mbbsEmuCpuRegisters.EAX);
 
             //Verify Flags
-            Assert.Equal(zeroFlagValue, mbbsEmuCpuRegisters.F.IsFlagSet((ushort)EnumFlags.ZF));
-            Assert.Equal(signFlagValue, mbbsEmuCpuRegisters.F.IsFlagSet((ushort)EnumFlags.SF));
-            Assert.Equal(overflowFlagValue, mbbsEmuCpuRegisters.F.IsFlagSet((ushort)EnumFlags.OF));
-            Assert.Equal(carryFlagValue, mbbsEmuCpuRegisters.F.IsFlagSet((ushort)EnumFlags.CF));
+            Assert.Equal(zeroFlagValue, mbbsEmuCpuRegisters.ZeroFlag);
+            Assert.Equal(signFlagValue, mbbsEmuCpuRegisters.SignFlag);
+            Assert.Equal(overflowFlagValue, mbbsEmuCpuRegisters.OverflowFlag);
+            Assert.Equal(carryFlagValue, mbbsEmuCpuRegisters.CarryFlag);
         }
     }
 }
