+++ MBBSEmu/Logging/LoggerExtensions.cs
@@ -73,11 +73,11 @@ namespace MBBSEmu.Logging
             output.Append($"SP={cpu.Registers.SP:X4}  ");
             output.AppendLine($"BP={cpu.Registers.BP:X4}");
 
-            output.Append(cpu.Registers.Carry ? "C" : "c");
+            output.Append(cpu.Registers.CarryFlag ? "C" : "c");
             //output.Append(cpu.Registers.Parity ? "P" : "p");
-            output.Append(cpu.Registers.Zero ? "Z" : "z");
-            output.Append(cpu.Registers.Sign ? "S" : "s");
-            output.Append(cpu.Registers.Overflow ? "O" : "o");
+            output.Append(cpu.Registers.ZeroFlag ? "Z" : "z");
+            output.Append(cpu.Registers.SignFlag ? "S" : "s");
+            output.Append(cpu.Registers.OverflowFlag ? "O" : "o");
 
             foreach (var line in output.ToString().Split("\r\n"))
             {
